requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.1
nltk==3.8.1
tensorflow==2.16.1
wordcloud==1.9.3
imbalanced-learn==0.12.3
joblib==1.4.2
jupyter==1.0.0
notebook==7.2.0
ipykernel==6.29.5
python-docx==1.1.0
nbformat==5.10.4